requirements.txt
gradio>=3.44.2
opencv-python>=4.5.5
numpy>=1.21.0
Pillow>=8.3.0

